fix: address fixes after review
