+++ change/@fluentui-react-headless-components-preview-3cc555dc-e346-4c29-bcd0-c8dfbf38544d.json
@@ -1,5 +1,5 @@
 {
-  "type": "minor",
+  "type": "patch",
   "comment": "feat: add headless Tags components",
   "packageName": "@fluentui/react-headless-components-preview",
   "email": "vgenaev@gmail.com",
