+++ packages/react-components/react-headless-components-preview/library/src/components/InteractionTag/InteractionTagSecondary/renderInteractionTagSecondary.ts
