+++ packages/react-components/react-headless-components-preview/library/src/components/InteractionTag/index.ts
@@ -1,9 +1,34 @@
 export { InteractionTag } from './InteractionTag';
 export { renderInteractionTag } from './renderInteractionTag';
-export { useInteractionTag, useInteractionTagContextValues } from './useInteractionTag';
+export { useInteractionTag } from './useInteractionTag';
+export { useInteractionTagContextValues } from './useInteractionTagContextValues';
 export type {
   InteractionTagSlots,
   InteractionTagProps,
   InteractionTagState,
   InteractionTagContextValues,
 } from './InteractionTag.types';
+
+export {
+  InteractionTagPrimary,
+  renderInteractionTagPrimary,
+  useInteractionTagPrimary,
+  useInteractionTagPrimaryContextValues,
+} from './InteractionTagPrimary';
+export type {
+  InteractionTagPrimarySlots,
+  InteractionTagPrimaryProps,
+  InteractionTagPrimaryState,
+  InteractionTagPrimaryContextValues,
+} from './InteractionTagPrimary';
+
+export {
+  InteractionTagSecondary,
+  renderInteractionTagSecondary,
+  useInteractionTagSecondary,
+} from './InteractionTagSecondary';
+export type {
+  InteractionTagSecondarySlots,
+  InteractionTagSecondaryProps,
+  InteractionTagSecondaryState,
+} from './InteractionTagSecondary';
