+++ packages/react-components/react-headless-components-preview/library/src/components/TagGroup/index.ts
@@ -1,4 +1,5 @@
 export { TagGroup } from './TagGroup';
 export { renderTagGroup } from './renderTagGroup';
-export { useTagGroup, useTagGroupContextValues } from './useTagGroup';
+export { useTagGroup } from './useTagGroup';
+export { useTagGroupContextValues } from './useTagGroupContextValues';
 export type { TagGroupSlots, TagGroupProps, TagGroupState, TagGroupContextValues } from './TagGroup.types';
